OSINT-War-Room-main - Copy/OSINT-War-Room-main/requirements.txt
2captcha-python==2.0.3
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anticaptchaofficial==1.0.69
anyio==4.12.1
apify_client==2.5.0
apify_shared==2.2.0
asarPy==1.0.1
asyncio==4.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
blis==1.3.3
capsolver==1.0.7
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromedriver-binary==74.0.3729.6.0
click==8.3.1
cloudpathlib==0.23.0
cloudscraper==1.2.71
colorama==0.4.6
curl_cffi==0.13.0
cymem==2.0.13
dateutils==0.6.12
et_xmlfile==2.0.0
fastapi==0.135.1
feedparser==6.0.12
frozendict==2.4.7
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
impit==0.11.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
multitasking==0.0.12
murmurhash==1.0.15
nltk==3.9.3
ntscraper==0.4.0
numpy==2.4.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.1
peewee==4.0.1
platformdirs==4.9.2
preshed==3.0.12
protobuf==7.34.0
py==1.11.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.2
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
setuptools==82.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.52.1
textblob==0.19.0
tqdm==4.67.3
trio==0.33.0
tweety-ns==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
vaderSentiment==3.3.2
wasabi==1.1.3
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.1
wsproto==1.3.2
yfinance==1.2.0
